+++ generators/ga.py
@@ -56,7 +56,8 @@ class Generator:
         self._evaluate()
 
     def save(self, folderpath):
-        shutil.rmtree(folderpath)
+        if os.path.exists(folderpath):
+            shutil.rmtree(folderpath)
         os.makedirs(folderpath)
         for i in range(len(self._chromosomes)):
             self._chromosomes[i].save(os.path.join(folderpath, f"chromsome_{i}.json"))
